requirements.txt
google-genai>=1.55.0
streamlit>=1.28.0

